+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- Create or modify events using `SentryEvent` objects and new SDK methods: `SentrySDK.capture_event(event)`, `SentrySDK.create_event()`, `SentrySDK.create_message_event()` ([#51](https://github.com/getsentry/sentry-godot/pull/51))
+- Create or modify events using `SentryEvent` objects and new SDK methods: `SentrySDK.create_event()`, `SentrySDK.capture_event(event)` ([#51](https://github.com/getsentry/sentry-godot/pull/51))
 
 ## 0.0.3